various changes

+++ core/src/main/java/hazelnut/core/config/HazelnutConfigBuilder.java
@@ -12,11 +12,11 @@ public interface HazelnutConfigBuilder {
 
     @NotNull HazelnutConfigBuilder policy(final @NotNull NoProcessorPolicy noProcessorPolicy);
 
-    @NotNull HazelnutConfigBuilder cacheExpiryRate(final int value, final @NotNull TimeUnit timeUnit);
+    @NotNull HazelnutConfigBuilder cacheExpiryRate(final long value, final @NotNull TimeUnit timeUnit);
 
-    @NotNull HazelnutConfigBuilder cacheHousekeeperRate(final int value, final @NotNull TimeUnit timeUnit);
+    @NotNull HazelnutConfigBuilder cacheHousekeeperRate(final long value, final @NotNull TimeUnit timeUnit);
 
-    @NotNull HazelnutConfigBuilder heartbeatRate(final int value, final @NotNull TimeUnit timeUnit);
+    @NotNull HazelnutConfigBuilder heartbeatRate(final long value, final @NotNull TimeUnit timeUnit);
 
     @NotNull HazelnutConfig build();
 }
